Moved SeqIfMainCharacterValere to its own file
