+++ route/evermist_island/mooncradle.py
@@ -10,16 +10,12 @@ from engine.seq import (
     SeqCheckpoint,
     SeqClimb,
     SeqHoldDirectionUntilLostControl,
-    SeqIf,
+    SeqIfMainCharacterValere,
     SeqInteract,
     SeqList,
     SeqMove,
     SeqSelectOption,
     SeqSkipUntilIdle,
-)
-from memory import (
-    PlayerPartyCharacter,
-    player_party_manager_handle,
 )
 
 logger = logging.getLogger(__name__)
@@ -118,16 +114,6 @@ class IntroMooncradle(SeqList):
                 ),
             ],
         )
-
-
-class SeqIfMainCharacterValere(SeqIf):
-    def condition(self: Self) -> bool:
-        leader = player_party_manager_handle().leader_character
-        if leader == PlayerPartyCharacter.Valere:
-            return True
-        if leader == PlayerPartyCharacter.Zale:
-            return False
-        return None
 
 
 class LoomsToCenter(SeqIfMainCharacterValere):
